feat(studio): relocate read replica modules under Infrastructure
Move RR form/row/modals into Settings/Infrastructure, extract REPLICA_STATUS, and add path helpers. Behaviour and URLs unchanged.
