+++ apps/studio/components/interfaces/Database/Replication/Replication.constants.ts
@@ -1,7 +1,3 @@
-import { components } from 'api-types'
-
-import { PROJECT_STATUS } from '@/lib/constants'
-
 export const STATUS_REFRESH_FREQUENCY_MS: number = 10000 // 10 seconds
 
 export enum PipelineStatusName {
@@ -13,10 +9,5 @@ export enum PipelineStatusName {
   UNKNOWN = 'unknown',
 }
 
-export const REPLICA_STATUS: {
-  [key: string]: components['schemas']['DatabaseStatusResponse']['status']
-} = {
-  ...PROJECT_STATUS,
-  INIT_READ_REPLICA: 'INIT_READ_REPLICA',
-  INIT_READ_REPLICA_FAILED: 'INIT_READ_REPLICA_FAILED',
-}
+/** @deprecated Import from Settings/Infrastructure/ReadReplicas/ReadReplicas.constants */
+export { REPLICA_STATUS } from '@/components/interfaces/Settings/Infrastructure/ReadReplicas/ReadReplicas.constants'
