+++ apps/studio/components/interfaces/Settings/Infrastructure/InfrastructureConfiguration/Edge.tsx
@@ -4,7 +4,7 @@ import { Loader2 } from 'lucide-react'
 import { Tooltip, TooltipContent, TooltipTrigger } from 'ui'
 
 import { EdgeData } from './InstanceConfiguration.constants'
-import { REPLICA_STATUS } from '@/components/interfaces/Database/Replication/Replication.constants'
+import { REPLICA_STATUS } from '@/components/interfaces/Settings/Infrastructure/ReadReplicas/ReadReplicas.constants'
 import { useReplicationLagQuery } from '@/data/read-replicas/replica-lag-query'
 import { formatDatabaseID } from '@/data/read-replicas/replicas.utils'
