+++ apps/studio/components/interfaces/Settings/Infrastructure/ReadReplicas/ReadReplicas.constants.ts
@@ -1,0 +1,11 @@
+import { components } from 'api-types'
+
+import { PROJECT_STATUS } from '@/lib/constants'
+
+export const REPLICA_STATUS: {
+  [key: string]: components['schemas']['DatabaseStatusResponse']['status']
+} = {
+  ...PROJECT_STATUS,
+  INIT_READ_REPLICA: 'INIT_READ_REPLICA',
+  INIT_READ_REPLICA_FAILED: 'INIT_READ_REPLICA_FAILED',
+}
